01/08/23

+++ src/main/java/sem_11/Ex001/CalculatorController.java
@@ -5,24 +5,24 @@ public class CalculatorController {
     private CalculatorView view;
     private CalculatorModel model;
 
-    public CalculatorController (CalculatorModel model, CalculatorView view){
+    public CalculatorController(CalculatorModel model, CalculatorView view) {
         this.model = model;
         this.view = view;
     }
 
-    public void perfomeOperation(String operation, int num1, int num2){
-        switch (operation){
+    public void perfomeOperation(String operation, double num1, double num2) {
+        switch (operation) {
             case "+":
                 model.add(num1, num2);
                 break;
             case "-":
-                model.subtract(num1,num2);
+                model.subtract(num1, num2);
                 break;
             case "*":
-                model.multiply(num1,num2);
+                model.multiply(num1, num2);
                 break;
             case "/":
-                model.divide(num1,num2);
+                model.divide(num1, num2);
                 break;
             default:
                 System.out.println("некорректная операция");
