+++ src/main/java/sem_11/Ex001/CalculatorModel.java
@@ -2,17 +2,17 @@ package sem_11.Ex001;
 
 public class CalculatorModel {
 
-    private int result;
+    private double result;
 
-    public void add(int num1, int num2) {
+    public void add(double num1, double num2) {
         result = num1 + num2;
     }
 
-    public void subtract (int num1, int num2){
+    public void subtract (double num1, double num2){
         result = num1 - num2;
     }
 
-    public void divide(int num1, int num2) {
+    public void divide(double num1, double num2) {
         if (num2 != 0) {
             result = num1 / num2;
         } else {
@@ -21,11 +21,11 @@ public class CalculatorModel {
         }
     }
 
-    public void multiply (int num1, int num2){
+    public void multiply (double num1, double num2){
         result = num1 * num2;
     }
 
-    public int getResult() {
+    public double getResult() {
         return result;
     }
